services/ark-api/ark-api/pyproject.toml
[project]
name = "ark-api"
version = "0.1.17"
description = ""
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "dpath>=2.2.0",
    "fastapi>=0.115.8",
    "kubernetes-asyncio>=33.3.0",
    "openai>=1.100.2",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.2",
    "uvicorn>=0.34.0",
    "httpx>=0.24.0",
    "pyhelm3>=0.4.0",
    "coverage>=7.10.4",
    "ark-sdk",
]

[tool.uv]
package = true

[tool.uv.sources]
ark-sdk = { path = "../../../out/ark-sdk/py-sdk/dist/ark_sdk-0.1.28-py3-none-any.whl" }

[tool.coverage.run]
data_file = "coverage/.coverage"

[tool.coverage.html]
directory = "coverage"

[tool.coverage.lcov]
output = "coverage/lcov.info"
